Make sure the OpenVPN log file is readable by non-root users
When OpenVPN creates it's log file it is only readable by root. When we
then want to collect this log file in the problem report we do not want
to be root. If OpenVPN on the other hand is told to log to a file that
already exists it doesn't change the permissions of that file. So here
we create the log file with permissions that work for us before OpenVPN
has the chance to create it itself.