Allow selecting relay in sole city of a country
Previously, when the code to allow selecting a relay in a city was
implemented, the code to expand a country missed a case where a country
could have a single city that had multiple relays. In that case, the
country wouldn't be expandable and the user wouldn't be able to select a
specific relay from that city. This fixes that issue by also allowing a
country to expand if it has only one city that has more than one relay.