Fix country not being selected on first run
Was happening because the first relay list fetch returned an empty list
and when attempting to find the relay item for the respective country
`null` was assigned to the `selectedRelayItem`. The setter for that
property lead the constraint to be reconfigured in the UI to be `Any`,
which meant no relay list item would be selected on the next relay list
update.