Check if returned tunnel is null
If the daemon tries to connect when our service has had it's permissions
revoked on Android, VpnServiceBuilder.establish() will return a null
poitner, which has to be checked for. By returning an error when this
happens, the daemon ends up in the error state where it's not blocking.
Whilst not ideal, this is preferrable to the crash.