Fix leak of connectivity subscriptions
The `LocationInfoCache` is created for every service connection, but the
`ConnectivityListener` it uses is reused from the service if the service
is always the same. Since there was unsubscribe command, the number of
listeners would increase every time the main app activity was started.

Each listener would then attempt to fetch the location if connectivity
was lost and then restored.