Change built `libwg.so` output path
Change the makefile to output to the path that will be used when
building the `mullvad-jni` crate, then copy that file to the path where
the APK build process will look for it. This is the reverse of how it
was previously handled.

The reason for doing is is that the APK library will be stripped, so
having that be the output of the makefile will always cause `make` to
rebuild it unnecessary if it's stripped in place. Therefore, we keep the
`make` output library unstripped, and copy a stripped version to the APK
path.