Remove old shared preferences clean-up code
On versions 2019.8-beta1 and older, the app stored the application
version information as shared preferences. On 2019.9-beta1 that changed
so that the information was handled by the Rust daemon, and some code
was added to clear those shared preferences so that users that update
from the old versions don't have any left over data.

This commit removes that migration code, because it's been more than a
year since that change has happened and none of those versions are
supported anymore. A user that updates from one of the old versions
before 2019.9-beta1 to a version from this commit onwards will have some
extra data stored in shared preferences (which should be less than 1
KiB) that won't be ever used. The only way to remove that data is to
manually clear the app storage, reinstall the app, or to upgrade to a
version from 2020 before updating to future versions.