Reduce output in extract-geo-data script
The large amounts of output in the `extract-geo-data.py` script makes it
difficult to spot real errors. Reducing the amount of output saying
whether or not a translation was found and which one is being translated
will make it easier to follow the output and make it easier to spot more
critical errors.