Avoid running in foreground when not connected
This change makes the Android service (MullvadVpnService)
only run in foreground when connected, compared to previously
when it also ran in foreground when the app/ui was in foreground.

This change reduces the amount of potential states after the
previous work of splitting the app/ui and service into separate
processes. It also makes the app/service better aligned with the
Android platform, as foreground services only should be used for
ongoing operations/work.