Fix custom dns toggle not working after resume
Fixes an issue with the custom dns toggle and server list not not
working after the app is resumed, due to the logic not being aware
of new service connections.

The issue was fixed by moving some responsibility from the adapter
to the fragment and making sure new service connections triggers new
data subscriptions to be set up.