Add Android dependency audit
Adds the OWASP Dependency-Check tool to the project to detect
potentially vulnerable dependencies.

Gradle is configured to skip the lintClassPath configuration, which
relies on many dependencies that has been flagged to have CVEs, as it's
related to the lint tooling rather than the project's compilation class
path. The alternative would be to suppress specific CVEs, however that
could potentially result in suppressed CVEs in project compilation
class path.