Fix notification update when vpn service killed
Fixes an annoying behavior where the vpn state notification is created
or updated just before the vpn service is killed. For example, if the
tunnel is disconnected and the user swipes away the "Unsecured"
notification, a new "Unsecured" notification (without the possibility to
connect) would appear after a delay (~1-3min).

This fix ensures that the notification isn't updated when the service is
killed.