Propagate selected obfuscator hostname to frontends
The obfuscator hostname should be propagated to the frontends, which
implies that it should be stored for later querying in the daemon. Since
all other relays were being kept as fields in the daemon god-struct, I
factored them out into a single enum containing all the individual
relays that could be selected.