Amend review comments reg naming and correctness
Amend review comments in regards to variable naming, tag  being assigned
inside of a loop and there not being a separate variable for the cleaned
version of tag. The cleaned version of the tag is also moved out of the
loop and is later checked to not be empty since semver does not allow an
empty string after '+'.