Move $cleaned_tag into for loop & always define it
Move the $cleaned_tag variable into the for loop scope in order to not
clutter up global variables and always define it either as the cleaned
tag or as "" in order to avoid crashing when it is undefined.

The purpose of using the cleaned tag inside of the `if [[ -n $cleaned_tag
]]` is to make sure that if the tag contained only disallowed variables
then it should not be used, so if the resulting cleaned tag is empty
then we do not append a `+`.