Create verification script and github action
Create a script which verifies that a set of "locked down" files are not
changed in commits that have not been signed.
Create a separate file that specifies which files are locked down and
another file which specifies all of the public key that are considered
trusted.

Create a github workflow that runs the script in the CI.