Fix a bug related to automatic MTU detection
A panic would happen when a route node had both an IP and a device and
a wireguard tunnel would start up. This has been patched and the panic
has been turned into an error that will only log a failure and cause the
default MTU to be statically determined.