Rely on device state for logout
Prior to the device management, the only way to logout was via the
account view. However, since the logic for handling device revocation is
similar to device logout, it can now be handled more or less the same as
part of the device state listener in the activity. This means that the
account view only has to notify the daemon about the logout which will
trigger appropriate transitions.