Fix newly introduced clippy warnings
Most significantly included making the TunnelArgs struct bigger by
including most of the parameters that are used in the call-chain for
Wireguard tunnels. This makes the struct less arbitrary and keeps down
the amount of arguments to a manageable level.

Fix `is_offline` lint complaint
Nightly clippy (0.1.63) does not complain about this however stable
(0.1.61) does and as such we'll change the name it. The mutability did not seem to
be necessary but we agreed the name change is good anyway.