Promote service to fg when actions provided
As of Android 8 (API level 26), the system has a strict timer where the
service must run `startForeground` within seconds after being started as
foreground (`startForegroundService`).

To mitigate cases where the service has been too slow to reach the
target state, for instance when the tile service starts the vpn service
as foreground with the connect action, this change aims to always
promote the service to foreground when connect/disconnect actions are
provided. This solution relies on the service demoting itself to from
foreground in cases where it shouldn't run as foreground.