Allow for migrating windows users to WG
Use a potential value between 0 and 1 provided by the API version check
to decide if the Windows client should use WireGuard or OpenVpn.
This decision is then persisted in the settings in the form of a random
value between 0 and 1 where if the value is higher than the value
provided by the API the client will use OpenVpn. If there is no value
provided by the API then the client will use WireGuard as the migration
has concluded.