Update tests to use OpenVpn and fix windows test
Use OpenVpn as the default in the tests since that will emulate the old
behavior that those tests tested for the best. This only impacts windows
and in the case where it impacts windows it will behave as expected when
using OpenVpn as the default.
When the behavior that ought to be tested is the new behavior then new
tests need to be written.

Fix some test compile errors on windows.