Update naming and remove comments
It would be too complicated to make the migration code be windows only.
It will not actually matter because the part which uses the final
default_tunnel_type that is feature gated to only windows.
Remove all the comments which suggest making features windows only.

This commit also updates some of the naming in order to make it more
clear.
default_tunnel_protocol -> default_tunnel_type
threshold_wg_default -> wg_migration_threshold