Make relay-selector respect configs tunnel default
Previously relay-selector did not respect the configs tunnel default
unless target_os was windows. Now it will always respect the provided
default. Make sure that the default created config always has wireguard
as a default except for on windows where it will calculate the default
using the provided wg_migration_threshold.

Change all the relay-selector tests to default to WireGuard unless they
are feature flagged for windows in which case we default to OpenVPN.