Make dispatch workflow work with verify script
Dispatch workflow could previously not run the verification script
correctly since it could only get the appropriate commit depth when a
pull request was made. Change this so that a dispatched workflow has a
input for the commit depth which defaults to 50. This should make it
play ball for testing purposes which is when we need dispatch workflow
to function.