Move adaptive presentation delegate from scene delegate into its own class.
Also fix issues in former implementation where transitionCoordinator.containerView
used to return UIWindow when changing to split screen.

New implementation makes a manual lookup of UITransitionView to ensure that
the settings cog is added into the right container view.

In the future this better be replaced with UISheetPresentationController subclass
but for now we'll keep adding the settings cog manually within the adaptivity
call.