Print `mullvad-app` version even if `mullvad-daemon` is down
Previously, if `mullvad-cli` could not establish a connection to
`mullvad-daemon` for any reason, running `mullvad version` would not
print any version information whatsoever. Now we do print the current
build version of `mullvad-app`, which may or may not be in sync with the
installed version of `mullvad-daemon`.

We use the `mullvad-version` crate to figure out an appropriate response
in case `mullvad-daemon` is down. In this case we know that we failed
creating a proxy client, so we add that piece of information to the
error context.