Allow dev builds to override location checking endpoint
If `mullvad-daemon` was built with the `api-override` feature flag, the
new environment variable `MULLVAD_CONNCHECK_HOST` may be used to supply
a different hostname for the location checking endpoint than the
standard `am.i.mullvad.net`.

It is equivalent to replacing

`curl https://ipv4.am.i.mullvad.net/json`

with

`curl https://ipv4.{$MULLVAD_LOCATION_HOST}/json`

Note that this change applies to both IPv4 and IPv6 endpoints.

Log warning if `MULLVAD_CONNCHECK_HOST` is used in release builds. The
variable will have no effect if `mullvad-daemon` was built in release
mode, but it is good to let the user know this if he/she thinks
otherwise.