Verify that important files have not been accidentally removed
Verify that we do not remove any file we deem important. If any such
file is removed, they should also be removed from the list this CI job
uses, and as such we raise a flag.

Update the CI job to also check that all files in the list of important
files still exists in the working directory/at the HEAD where it's being
executed from.