Improve UX of CLI when user connects while not logged in
If the user is *not* logged in to an active account, we will from now on
issue a warning if they try to connect to a relay using `mullvad-cli`.

Previously, we would silently just start blocking all internet traffic.
This is well and good, as this is mandated by the state machine model.
The huge issue is that the user was unaware about the current status of
their connection without running further CLI commands or checking the
GUI. As this is to be considered a frustrating edge case, we might as
well try to help the user, since they can't look for answers on the
world wide web.