Fix CI verification script
The verification script running on github would complain that the first
commit of a PR changed all of the locked down files if this commit was
unsigned. The reason was that in our github enviornment we don't fetch
the entire git history but instead only the needed commits (so called
shallow repository). The entire history is therefore folded into the
first commit in the shallow repo. Since this history contains
modifications of locked down files and since this commit is not signed
the script will think this commit modified files its not allowed to.
This fix deepens the shallow repo by 1 after fetching.
It also clarifies a print in verification script to be more descriptive.