Make `mullvad relay set location` smarter
The `set location` command now takes a hostname and figures the country
and city out. This is identical to how the (now deprecated) `mullvad
relay set hostname` used to work.

The `set location` command will try to resolve country code first, but
if that fails we now fall back to trying to resolve a relay by hostname
first.

Update the help message (long & short) to cover this new use case.