Do not consider removing a revoked device to be an error.
Sometimes when removing a device when uninstalling the app, a `404` is
returned if the device has already been revoked from the account.
However, removing a non-existent device should not be considered an
error if the end goal is to not have the device on the account. That
goal has clearly already been achieved.