Fix hashing of obfuscation protocols should not consider enabled state
What does it mean to compare two different obfuscation protocols? A logical
semantic would be to compare the data related to networking (such as ip
addresses & port numbers) in order to
determine if one SOCKS5-proxy is equal to another SOCKS5-proxy. But
since the introduction of the `enabled` member, which signals if a given
instance of any proxy protocol should be used for accessing the mullvad
API, this temporal value broke that semantic which resulted in equality
comparisons and hashing of two similarly-looking instances to return
unexpected results.

This change stabilizes the above semantics by hand-rolling `Hash` for
all the supported obfuscation protocols.