Add new module for defining rpc-only types in `mullvad-management-interface`
This is mostly a prototype of a layout that I found logical while moving
the type `ApiAccessMethodUpdate` (which only exist at the gRPC layer). I
did not find that any other such type had been created in
`mullvad-management-interface` before, so I either had to

- fit the type definition into an existing module, such as `conversions`
- create a new module

The former option broke the pre-existing invariant that only conversion
code (mostly `impl From<This> for That`) lived in the `conversions`
module, so defining data types there felt unnatural.