Improve printing, rename and comment things
Printing is better and refactored to be shared with access methods,
renamed and made a clear distinction between what is called a "Custom
Bridge" and what is called a "Custom Proxy", the former is the specific
feature, the latter is the shared type between custom bridges and api
access methods.

Create doc comments.