Properly pass arguments to Cargo in `build-apk.sh`
The `build-apk.sh` script broke in b6e0545e1b5ae2fbcb6655bd4d28739cf9f943d6
due to quoting a variable containing a string with space seperated
arguments. While it is best practice to quote variables to prevent word
splitting, word splitting was the behavior that we wanted in this
particular case.

This commit fixes this issue by building an array of arguments instead,
which when expanded will pass along the arguments to `cargo build` properly
as multiple strings. The script now works, and `shellcheck` is happy!