Build `wireguard-go` if needed
`talpid-wireguard` links to `wireguard-go` on all platforms except
Windows. Building `talpid-wireguard` was a bit flaky due to making an
assumption that `wireguard-go` has been built in a previous build
step (outside of Cargo).

This commit adds a check for `wireguard-go` when building the
`talpid-wireguard` crate - if the static library does not exists, we
will call the script for building `wireguard-go` before compiling
`talpid-wireguard`.