Run more GH jobs in parallell to improve speed
This is achieved by no longer re-using build
results between steps and instead running
some of the common gradle tasks in a redundant
parallall way. Also makes use of a third-party
action called wait-for-jobs to pause at a step
in the middle of a job and awaiting artifacts
from other jobs.