Reorder actions to not hide important failure
The mristin/opinionated-commit-message action can sometimes fail on
things we don't consider blocking/hard errors. Such as a correct
imperative verb that does not happen to be in the action's wordlist.
In these cases the job is aborted and the whitespace checking step is
never executed, leading to a larger risk of accidentally merging invalid
whitespaces