Ignore android vulnerabilities reported by osv-scanner
Temporarily ignoring all reported android vulnerabilites
with a one month deadline for osv-scanner that we are
adding to our suite of tools. The reason for this is that
we plan to examine the vulnerabilites and bootstrap this
file with proper ignore reasons (or address by bumping
dependencies).

Also worth mentioning that we're already using the OWASP
Dependency-Check tool for the android code base as of
before.