Improve variable names, comments and error messages
Briefly explain the usage.
Make it clearer that the program does not work with an IPv6.
Remove the somewhat confusing usage of `trim`. The IP is not trimmed
either and it would only protect from a purposefully wrong argument
(With quoted or escaped spaces.)
Explain what the boolean parameters stand for.
Explain why the PSK has to be printed with `fmt::Debug`.