Only accept an IPv4 as valid first argument
While `request_ephemeral_peer` accepts a generic `IpAddr`, the
function `new_client` used within forces an `Ipv4Addr`.

In the long run it would be best to change the library to support IPv6
(provided that the tunnel configuration service listens to an IPv6) or
to change the signature of `request_ephemeral_peer` to expect an IPv4.