Refactor test cleanup logic
Cleanup is now done BEFORE tests are run and takes care of resetting
the daemon state more thoroughly. The daemon will now always be
installed, logged in and disconnected with all settings reset before
the next test. Tests are therefore not able to depend on the previous
test leaving the test-runner in a certain state and must instead take
care of setting up their own state themselves.

`test_upgrade_app` gets special treatment to be able to run before
the new app version is automatically installed.