Remove cargo dependency when running e2e tests
This commit enables the usage of the dist/ directory, and also adds
mullvad-version to it so that test-by-version.sh can operate without
rust installed at all.

To make use of predefined binaries in a separate directory, refer to
that directory by using the env var TEST_DIST_DIR=<dir path> and the
binaries will be used if they can be found there. If TEST_DIST_DIR is
specified, all of the following binaries need to be there:
- connection-checker
- mullvad-version
- test-manager
- test-runner

Also added a /dev/null redirect of a cd output so that one's able to use
CDPATH while running the tests.