Fix IP-override feature indicator
It was trigger by any overrides existing in the settings, not by the
current endpoint being overridden.

Add flag to `Relay` to specify if its IPv4 and/or IPv6 has been
overridden and use that in combination with the endpoint IP version to
derive if the current connection is overridden.