Add top level `permissions` to all Github Actions workflows
The default permission on the repository is already set to read only. So
in practice this makes no difference. But this makes that more explicit,
and less relying on the repository being correctly configured.

This also makes security scanning tools such as OpenSSF scorecard
happier about the overall security of our repository.