Upgrade `tonic` and `prost`
Upgrade `tonic` to version 0.12, as a prerequisite
to later on upgrading `hyper`.

As of version 1, `hyper` no longer uses tokios `AsyncWriter`
and `AsyncReader` traits, instead defining its own versions,
see <https://github.com/hyperium/hyper/issues/3110>. As tonic
`0.12` is updated to use the `hyper 1.0` ecosystem, it changed
some of its trait-bounds to the new `hyper` traits. The `hyper-utils`
crate provides the wrapper `TokioIo`, which converts between the two.

`prost` had to be upgraded as well, for compatibility.