Add dedicated "known issues" document
Moves some known issues out of the security document.
Keeps all known issues in one place.
Allows the security document to be more focused on the app threat model
and how it intend to solve that threat model. The security document
is a more static document that should rarely change. While the known
issues document will grow and shrink as new special cases are found and
fixed.