Make smart routing take precedence over multihop
If both multihop and DAITA (with smart routing) is enabled, a
DAITA-compatible entry relay will be selected. This implies that if the
user has selected an entry relay which is not DAITA-enabled, the relay
selector will override this choice and force a DAITA-enabled relay as
entry. If smart routing is disabled in this case, the user's selected
entry will always be selected, even if this means that the user will end
up in a blocked state.