Disable Apple services workaroudns for unaffected macOS versions
Do not apply redirect rules for DNS requests to port 53 to loopback
on macOS versions that do not need the apple services NAT-redirect workaround.
This effectively reverts to the old behaviour were a local DNS resolver is used
only in the blocked and connecting states for macOS versions *not* in the version
range 14.6 <= version < 15.1.

Revert the nat redirect rules that were added to force traffic through
the tunnel interface. This hack is no longer needed since it was fixed
by apple in macOS 15.1.