Add pinned Rust version via rust-toolchain.toml
As part of enabling reproducible builds we need to build a given commit
with a pinned Rust version.

This commit adds a minimal rust-toolchain.toml file that specifies this
version (this Rust version is enforced by rustup when invoking a cargo
command).

In the container we extract the version that is specified in
rust-toolchain.toml to ensure that the container is pre-built with the
correct Rust version (so that when buildling in the container we do not
need to download the right Rust version).