Finish attempt at cancellation from the rust side.
Now `mullvad_api_get_addresses` returns a cancel handle. It should be
wrapped in a class on the swift side so that it is deallocated
appropriately. But calling it should allow for cancelling any in-flight
requests. It doesn't wait for the task to actually stop - but it we can
make it do it.