Add setup-rust script and git hook
The setup-rust helper script installs the default targets and
components that each platform needs with rustup.

The setup-rust-post-checkout script is a git hook (needs to be
copied to .git/hooks/post-checkout, which can be done manually or
by running `setup-rust install-hook`) that detects when
the rust-toolchain.toml file has changed and installs the default
targets/components (and/or extra targets/components that the user
has specified in the hook script).