Add `setup-rust` script and git hook
The `setup-rust` helper script uses Rustup to install the default
targets and components that each platform needs.

The `setup-rust-post-checkout` script is a git hook (which needs to be
copied to .git/hooks/post-checkout, which can be done manually or
by running `setup-rust install-hook`) that detects when
the `rust-toolchain.toml` file has changed and installs the default
targets/components for the platform that the user has specified via
the MULLVAD_SETUP_PLATFORM env variable.