Format workspace
The 2024 edition contains new formatting rules.
All of these are not compatible with the 2021
formatting style, but most of them are.

To change the formatting to be as close to the
new style as possible while remaining compatible
with the 2021 edition, I first ran
`cargo +beta fmt` with the edition set to 2024,
then, with the edition set to 2021, I ran
`cargo fmt`.