Change script file to load in index.html after vite migration
With the migration to vite we do not need to
pre-bundle or transpile code before referencing
it in the index file, as vite will resolve and
perform the necessary actions for us, i.e.
transpile the typescript and bundle it.

The global exports variable was only used for
the gulp bundle and has therefore also been removed.