Set postinstall script to build typescript lib
We want the typescript files to be available
after the npm installation has completed.[1]

Previously we only built the typescript files in
packages when they would be used on the platform
we are building for.

However, until the typescript files have been
generated we can't use imports in the mullvad-vpn
package to the other packages, such as
win-shortcuts.

If we use imports from these packages then the
builds will fail on platforms where these
packages are not used, as they haven't been built
before we transpile typescript into javascript.

The build failure occurs as typescript will try
to transpile all files in the directories
specified in tsconfig.json, regardless if the
file will actually be used for the platform or
not.

[1] https://linear.app/mullvad/issue/DES-1588/switch-to-other-build-system#comment-3546e3b9