Ensure items in custom lists respect disabledLocation
When a custom list is selected as Entry or Exit
and only contains a single relay, then that relay
should be disabled on the opposite view, which is
what already happens for its custom list parent.

When a single relay is selected as Entry or Exit
then that relay should be disabled in the
oppositve view.