WIP REMOVE Adjust button to ensure all styles pass through styled components
Right now we're seeing some flakyness in our E2E
tests because Playwright's Locators incorrectly
matches all buttons because they share the exact
same class names, which is what Playwright uses
in the end to target the DOM element.

This happens because styled-components generates
class names for a styled component based on the
props passed to it, and it seems like when the
Button component was added in this PR[1] we have
moved much of the styling inside the style
object, which styled-components does not use when
it generates the class names. The end result is
that all buttons which aren't disabled share the
exact same class names.

[1] https://github.com/mullvad/mullvadvpn-app/pull/7312