Use regular default route for the tunnel interface on Windows instead of /1 routes
This is done for two reasons:

1. This mitigates an issue in our fork of wireguard-nt, which intentionally allows routes
back to the tunnel interface. The fork explicitly disallows this only for routes with a prefix 0,
which means that the /1 routes are not exempted. This can result in an infinite routing loop if
the non-tunnel route to the relay is removed (e.g., if the default interface or its routes
disappear).

2. This simplifies the code and routes. The /1 routes are unnecessary since we're
setting the metric on the default route and interface to the lowest value anyway, so the
OS should always prefer the tunnel default route. Even if it doesn't, the firewall will
prevent leaks.