Add override for vite
A CVE exists in the 6.2.0 version of vite.
We have already upgraded the dependency on vite
to 6.2.3 where this CVE has been fixed.

However, npm resolves the @vitejs/plugin-react
package's dependency on vite to 6.2.0 even after
we have upgraded it for some reason.

To ensure the old 6.2.0 version is not resolved
we add an override to 6.2.3 until we can bump the
@vitejs/plugin-react package.