Disable pre-generation of PQ keys on Android
By generating 2 PQ keys on start we slow the start up of the android app
with out about 20%. On android we don't have the daemon running
continously and by default PQ nor Multihop is enabled, thus all our
users doing a cold start of the app will be 20% slower. Ideally the
solution would be to conditionally change the buffer size depending on
what features are active, e.g PQ gives buffer of 1, PQ and Multihop
gives a buffer of 2.