Update "Upgrade app" button to condition action based on platform
The action should be to open the download url on Linux
and to go to the App Upgrade route for Windows and Mac OS.

The "Upgrade app" button should only be disabled if the app
is offline when on Linux. Otherwise we should always be able
to navigate to the App Upgrade route.