Implement manage devices screen
This commit contains changes to the confirm dialogs because
the API for InfoConfirmationDialog was changed so that there are now
two variants, one that takes ResultBackNavigator<T> and one that takes
ResultBackNavigator<Confirmed>, where Confirmed signifies that the user
pressed the confirm button (instead of using a boolean for this). This
means that if the result is typed Confirmed the result can only be
Canceled or Confirmed (instead of the previous 3 cases: Canceled,
true, false).