Use builder pattern to construct `quic::Settings`
Make fields of `quic::Settings` private and expose constructors/builder
fns to configure a `quic::Settings` struct. A previous commit that added
a public, optional field to the settings struct broke a caller (iOS),
which is suboptimal.

Add a new token type: `quic::AuthToken`. This will perform "all"
necessary validation that was previously done ad-hoc when the token was
to be transformed into an authorization header. It also helps to not
fuck up the values passed to `quic::Settings::new`, since it was
previously Stringly-typed.