Set relay to current country on first start
Sets the default relay selection to the current country (as determined
by am.i.mullvad.net).

In non-release builds of the Android app we do not bundle a relay list
in the APK, and the relay list is fetched when the user logs in.

So one of the following can happen:

1. Geolocation request returns, we have a relay list.
2. Geolocation request returns, we do not yet have a relay list.
3. Relay list request returns, we have a geolocation.
4. Relay list request returns, we do not have a geolocation.

In 1. and 3. we can update the default locaiton. In 2. we have to wait
until the relay list is fetched from the api until we can update the
default location. 4. is unlikely to happen but could happen if
am.i.mullvad is down.