Improve account data fetching
Add a new account data fetch every time a user enters the connect screen
This is limited to at a maximum one fetch every minute.

Add a check that the user is still logged in to the same account
before updating the account data cache.

Change account fetching behavior in the account screen to
fetch on every enter instead of init.