Disable linting on React hooks which invokes a useEffectEvent callback
Currently React has not released their useEffectEvent hook[1] in a
stable version and we instead rely on our own implementation of the
hook to achieve the same functionality.

Unfortunately however, the react plugin for eslint does not recognize
useEffectEvent and that it can be omitted from a hook's dependency
array, hence we have to disable that rule until this issue can be
addressed.

Potential fixes will be discussed in the following linear issue:
https://linear.app/mullvad/issue/DES-2381

[1] https://react.dev/reference/react/experimental_useEffectEvent