Set the default location set flag at login
If the user has managed to log in we unconditionally set
update_default_location = false. This is done to prevent an edge case
where the geolocation request completes after the user has logged in and
then the default location could change while the user is in the process
of clicking the select location/connect buttons.