Add inital baseline profile generation
To improve startup performance this
PR adds a baseline profile generation module in test/baselineprofile.

The baseline profile plugin requires Junit4 so that is also added as a
dependency.

A baseline-prof.txt was also generated by running
`./gradlew generatePlayProdReleaseBaselineProfile` and checked in.

The tests that generate the baselineprofile currently only start the app
and accepts the privacy policy. This should be improved later on to
improve the startup performance.