Add access method update_builtin function
This access method's update function was rewritten to introduce
validation before updating the relevant access method, and therefore
when the function is used the potential error must be handled.

However, we still need to be able to update an access method
without any error handling, such as setting the Direct access method
as enabled, and that's where this new update_builtin function comes in.