Change TS moduleResolution to bundler and update TS output to es2022
When the @vitejs/plugin-react package was updated it was no longer
compatible with the TS config's `moduleResolution` setting, as such the
TS config had to be updated.

`moduleResolution` has to changed to `bundler`, and because of that we
can no longer target `commonjs` as the output format. As such we also
had to bump the `module` config option to `es2022`.

With these changes in mind it made sense to also bump the TS target to
`es2022` from `es2021`. The emitted code built by typescript is now
a lot more modern than in previous builds.