Do not set max_udp_payload_size in masque
There is no point in setting max_udp_payload_size to limit the size
of packets we can receive, because because the relay will do path-MTU
detection anyway.

There is also no point in limiting the size of packets that we send,
to max_udp_payload_size, since quinn does path-MTU detection which
we must respect regardless.

We were basing the value of max_udp_payload_size on the link-MTU
(not path-MTU), and not accounting for whether that value should
change over the lifetime of the connection. The link-MTU is mostly
uninteresting, because it is always `>= 1500` unless the user has
changed it, which most people don't.

We do technically still set `max_udp_payload_size`, because the
server requires it, but we always set it to the default value of
`1500 - IP_HEADER`