Do not set max_udp_payload_size in masque
There is no point in enforcing max_udp_payload_size because QUIC
already does path-MTU detection, which we need to respect anyway.

We were basing the value of max_udp_payload_size on the link-MTU
(not path-MTU), and not accounting for whether that value should
change over the lifetime of the connection. The link-MTU is mostly
uninteresting, because it is always `>= 1500` unless the user has
changed it, which most people don't.

We do technically still set `max_udp_payload_size`, because the
server requires it, but we always set it to the default value of
`1500 - IP_HEADER`