Fix app not launched by Linux split tunneling warning dialog button
The bug occurred because when the Launch button was clicked it called
the returned function from the useLaunchApplication hook. However, that
hook should only be used when the application list item is clicked,
since it is what opens the warning dialog in the first place.

Instead, the Launch button should call the onSelect function from the
LinuxApplicationRow context directly, and then hide the warning dialog.

This bug was introduced when the Split tunnel component was refactored.
See: https://github.com/mullvad/mullvadvpn-app/pull/8708