Add support for multihop entry filters in daemon
In the upcoming re-design of select location, separate sets of filters
can now be picked for the entry and the exit relays. This commit adds
support for that in the relay selector.

In order to not affect the current behavior of the desktop and Android
apps before the new UI is implemented, the entry filters are set to the
same as the exit filters when the relay settings are updated via gRPC.