WIP: at least it kinda works
TODO: Move the storekit2 transaction listener into the payment manager.
This will allow us to keep track of payments made in-band, we can share
a mutable map of the payment IDs so that we don't upload them twice. Or
maybe we do end up just having the listener inform all the listeners
whenever it receives the transaction via `Transaction.updates`.