Clear OOT notification on login/revoke screens
There are cases when a user may get to the login or device revoked
screens without their previously scheduled account expiry notification
being cancelled.

This can happen if for some reason the dameon has not sent the logged
out/revoked event, for example if the account is logged out via the
website while the app is not running.

This PR cancels the notification when the user gets to the login or
device revoked screens.