Log Kermit logs to disk
The FileLogWriter class logs all Kermit logs from the app to a log file
saved in the data/app_logs directory.

It will create one log file per day and supports log rotation based on
the number of days stored or the total size of all the log files in
data/app_logs.

These logs are appended to the end of the problem report.