Update v11 settings migration to do partial migration on Android
Updates the v11 to v12 settings migration to not migrate the setting
block_when_disconnected on Android, as that key does not and has never
existed on Android. The absence of that key caused the migration to
fail and prevented the later settings migrations from running.

To fix this issue the block_when_disconnected migration is only run on
platforms other than Android, i.e. Linux, macOS and Windows.