Remove bridging and fix networkd config
Due to rudimentary ambitions of running the WiFi network for the test
router off of the built-in WiFi chipset, there was a bridge interface
created that was now useless. This complicated matters a bit, so I
removed it. But the underlying issue is that the LAN interface was being
configured by the wrong networkd network configuration file since there
exists something along the lines of `99-ethernet-default-dhcp.network`
which supersedes the nicely named `lan.network` configuration. Thus, the
ultimate fix is to name our network config files `1-lan` and `1-wan`.

These changes also add git to the environment of the router to make
fixing this next time way easier.