Update user account expiry before showing notification
Since it is possible for users to update their account time
outside of the app the account will expire soon notification
can refer to an outdated expiry date.

To avoid this we will always start a background work to check
the latest expiry date before sending a notification.

If it is not possible to update the expiry date it will not show
a notification.